Improved solution finding a tiny bit by not auto-picking the first valid role of a player
In case there are multiple options a player can play to fill a role (e.g. ctank/hfbtank) it used to always pick the first option, which is determined by the roles.csv.
However this sometimes leads to it not being to create a solution because not all possible paths are selected (in the example above the hfbtank will never be selected).
Its not a proper solution as in includes randomization into something that is logical and structured, but the way "setPlayer" works by using "leftover" counters it would assign a player to multiple roles resulting in invalid squad results. Can be either resolved by always computing the amount of something that is still required, or re-upping the amount required if a role that is being checked doesn't turn into a valid solution.